backend/requirements.txt
fastapi[all]
python-dotenv
zeep
pydantic
pydantic-settings
sqlalchemy
pymysql
